Move helper types into their own namespace.
