+++ src/NodaTime.Web/Helpers/ProtobufConversions.cs
@@ -2,11 +2,8 @@
 // Use of this source code is governed by the Apache License 2.0,
 // as found in the LICENSE.txt file.using System;
 using Google.Protobuf.WellKnownTypes;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace NodaTime.Web
+namespace NodaTime.Web.Helpers
 {
     // This will be in its own package one day...
     public static class ProtobufConversions
