+++ src/NodaTime.Web/Models/GoogleStorageBenchmarkRepository.cs
@@ -6,6 +6,7 @@ using Google.Cloud.Storage.V1;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Logging;
 using NodaTime.Benchmarks;
+using NodaTime.Web.Helpers;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
