+++ src/NodaTime.Web/ViewModels/CompareTypesViewModel.cs
@@ -3,6 +3,7 @@
 // as found in the LICENSE.txt file.
 using MoreLinq;
 using NodaTime.Benchmarks;
+using NodaTime.Web.Helpers;
 using System;
 using System.Collections.Generic;
 using System.Linq;
