+++ src/NodaTime.Web/Views/_ViewImports.cshtml
@@ -1,3 +1,4 @@
 ﻿@using NodaTime.Web
+@using NodaTime.Web.Helpers
 @using NodaTime.Web.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
